backend/requirements.txt
asgiref
cachelib
autobahn
attrs
certifi
cffi
channels
charset-normalizer
click
colorama
constantly
drf_msal_jwtdaphne
distlib
django-filter
django-jwt-auth
djangorestframework
filelock
hyperlink
idna
incremental
itsdangerous
platformdirs
pyasn1
pyasn1-modules
pycparser
python-dateutil
pytz
requests
service-identity
txaio
typing-extensions
tzdata
urllib3
six
pymemcache
# msalpymemcache`
mysqlclient
djangorestframework-jwt
Automat
Django
Jinja2
Markdown
MarkupSafe
PyJWT
pyOpenSSL
PyYAML
Werkzeug
msal
drf_msal_jwt